% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxplot.R
\name{npplot}
\alias{npplot}
\title{Nonparametric and semiparametric Cox regression model.}
\usage{
npplot(temp, CIlevel = 0.95)
}
\arguments{
\item{temp}{Estimation result from function 'npcox' or 'spcox'.}

\item{CIlevel}{The default confidence level stands at 0.95.}
}
\value{
The plot of nonparametric coefficients function 'npcox' or 'spcox'.
}
\description{
Plot of time-varying coefficient estimated through function 'npcox' or 'spcox'.
}
\details{
This is some description of this function.

This is the plot function for
}
\examples{
data(pbc)
pbc = pbc[(pbc$time < 3000) & (pbc$time > 1500), ] 
Z   = pbc[,c("age","edema")]
colnames(Z) = c("age","edema")
del = pbc$status
tim = pbc$time
res = npcox(cva = Z,delta = del, obstime = tim, bandwidth = 500)
op  = par(mfrow = c(1,2))
npplot(res)
par(op)
}
