% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsimu.R
\name{npsimu}
\alias{npsimu}
\title{Simulation data generation of PH model with time-varying coefficients.}
\usage{
npsimu(n, cenpara = NULL)
}
\arguments{
\item{n}{The number of sample size, which can be self-determined.}

\item{cenpara}{Censoring parameter, which is supposed to be positive, for adjustment of censoring rate.}
}
\value{
a list that contain covariates, observed time and censoring indicator.
}
\description{
An example of data generation for nonparametric PH model.
}
\details{
This is some description of this function.

'np_simudata' is designed for PH model with time-varying coefficients, h(t) = h0(t)exp(b(t)'Z), generating the covariates, observed time and censoring indicator.
}
\examples{
data = npsimu(200)
}
